+++ CHANGES.md
@@ -37,6 +37,7 @@ Changes in CUPS v2.4.20 (YYYY-MM-DD)
 - Fixed potential buffer overflow in `cupsCopyDestConflicts` (Issue #1631)
 - Fixed backchannel parsing bug in `commandtops` filter (Issue #1637)
 - Fixed section parsing in the web help indexing code (Issue #1641)
+- Fixed media selection with a mix of PPD and IPP options (Issue #1651)
 - Fixed several issues reported by Coverity
 - Fixed case-sensitive PPD keyword comparisons when filtering keyword updates
   from filters.